requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
Django==4.2.13
djangorestframework==3.15.2
djongo==1.3.6
dnspython==2.6.1
idna==3.7
pymongo==3.12.1
python-decouple==3.8
pytz==2024.1
requests==2.32.3
sqlparse==0.5.0
typing_extensions==4.12.2
urllib3==2.2.2
